feat: 메인페이지 로그인/로그아웃 버튼 제거 및 마이페이지로 기능 이동
- HomeScreen에서 로그인/로그아웃 버튼 UI 제거
- MyPage에서 로그인 상태에 따라 로그인/로그아웃 버튼 분기 처리
- 로그인하지 않은 경우 로그인 버튼 노출 및 로그인 화면으로 이동
- 로그아웃 시 AsyncStorage 초기화 및 메인화면으로 리셋 이동 처리
